Change the upstream db into something more sensible
It now contains a link between upstream and the specfile repo, not the
other way around. We actually don't need the upstream link for a specfile
repo as the info is embedded in the specfile itself.

Also add a [copr] section capability so we can specify which copr to build
through copr-cli.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>